refactor wizard preview components
